Split settings form into tabbed components
