+++ src/components/SettingsFormBackgroundTab.svelte
