+++ src/components/SettingsFormSidebar.svelte
@@ -1,0 +1,11 @@
+<aside>
+  <label>
+    <input type="radio" name="sidebar" />
+    <span>Theme</span>
+  </label>
+
+  <label>
+    <input type="radio" name="sidebar" />
+    <span>Background</span>
+  </label>
+</aside>
